Archives: Less blog-like archive pages
As taxonomies are used extensively by news publishers to provide subject-specific areas the use of "Category:" and similar in the archive template title doesn't look great and makes it feel more blog-like than news publication.

This PR removes the prefix from taxonomy archives and improves the wording on date and author archives to be more descriptive.

Fixes #531